prettier and rename .mjs->.js
